Update foreman to work with command mapping
The current implementation does not honor command mappings. For example,
adding foreman to the list of `bundle_bins` used by the
capistrano-bundler gem should prefix the foreman execution with `bundle
exec`, but this does not happen in the current implementation.

This commit fixes the issue by passing in the command as an argument
array rather than a single string. This allows `SSHKit::Command.new` to
properly identify `foreman` as the command and apply mappings
accordingly.