Allow the upstart command to be customized
In my usecase, I am using [user
jobs](http://upstart.ubuntu.com/cookbook/#user-job) to control the
application. This allows me to avoid giving sudo access to the deploy
user; however, by default the deploy user does not have `/sbin` in the
`$PATH`. This change allows me to set `upstart_cmd` to `/sbin/initctl`
so the deploy user can start/stop the app. It also provides better
integration with the command mappings in SSHKit by using a single
command (`initctl`) and calling execute in a way that allows the command
to be properly identified.