Revert to using sudo over as(:root)
Using sudo should be safer in practice because it is easy to control
through the sudoers file. `as(:root)` requires unlimited access to the
root user, which is probably not what we want.

This reverts foreman and upstart back to using sudo, but the
implementation relies on SSHKit's command map prefixes. This eliminates
the need to branch in every task that may or may not need sudo.